# Add support for fallback from compute type f16 to f32