--- a/clients/gtest/auxiliary_gtest.cpp
+++ b/clients/gtest/auxiliary_gtest.cpp
@@ -27,7 +27,6 @@
 #include "hipblaslt_datatype2string.hpp"
 #include "hipblaslt_test.hpp"
 #include "testing_auxiliary.hpp"
-#include "type_dispatch.hpp"
 #include <cctype>
 #include <cstring>
 #include <type_traits>
@@ -39,47 +38,7 @@ namespace
     // aux
     // ----------------------------------------------------------------------------
 
-    // In the general case of <Ti, To, Tc>, these tests do not apply, and if this
-    // functor is called, an internal error message is generated. When converted
-    // to bool, this functor returns false.
-    template <typename TiA,
-              typename TiB  = TiA,
-              typename To   = TiB,
-              typename Tc   = To,
-              typename TciA = TiA,
-              typename TciB = TiB,
-              typename      = void>
-    struct aux_testing : hipblaslt_test_invalid
-    {
-    };
-
-    // When Ti = To = Tc != void, this test applies.
-    // When converted to bool, this functor returns true.
-    template <typename TiA, typename TiB, typename To, typename Tc, typename TciA, typename TciB>
-    struct aux_testing<
-        TiA,
-        TiB,
-        To,
-        Tc,
-        TciA,
-        TciB,
-        std::enable_if_t<
-            (std::is_same<TiA, hipblasLtHalf>{} && std::is_same<TiB, hipblasLtHalf>{})
-            || (std::is_same<TiA, hip_bfloat16>{} && std::is_same<TiB, hip_bfloat16>{})
-            || (std::is_same<TiA, float>{} && std::is_same<TiB, float>{})
-            || (std::is_same<TiA, hipblaslt_f8_fnuz>{} && std::is_same<TiB, hipblaslt_f8_fnuz>{})
-            || (std::is_same<TiA, hipblaslt_f8_fnuz>{} && std::is_same<TiB, hipblaslt_bf8_fnuz>{})
-            || (std::is_same<TiA, hipblaslt_bf8_fnuz>{} && std::is_same<TiB, hipblaslt_f8_fnuz>{})
-#ifdef ROCM_USE_FLOAT8
-            || (std::is_same<TiA, hipblaslt_f8>{} && std::is_same<TiB, hipblaslt_f8>{})
-            || (std::is_same<TiA, hipblaslt_f8>{} && std::is_same<TiB, hipblaslt_bf8>{})
-            || (std::is_same<TiA, hipblaslt_bf8>{} && std::is_same<TiB, hipblaslt_f8>{})
-#endif
-            || (std::is_same<TiA, double>{} && std::is_same<TiB, double>{})
-            || (std::is_same<TiA, hipblasLtInt8>{} && std::is_same<TiB, hipblasLtInt8>{})
-            || (std::is_same<TiA, hipblaslt_f8_fnuz>{} && std::is_same<TiB, hipblasLtHalf>{})
-            || (std::is_same<TiA, hipblasLtHalf>{} && std::is_same<TiB, hipblaslt_f8_fnuz>{})>>
-        : hipblaslt_test_valid
+    struct aux_testing: hipblaslt_test_valid
     {
         void operator()(const Arguments& arg)
         {
@@ -143,7 +102,7 @@ namespace
         // Filter for which types apply to this suite
         static bool type_filter(const Arguments& arg)
         {
-            return hipblaslt_simple_dispatch<type_filter_functor>(arg);
+            return type_filter_functor{}(arg);
         }
 
         // Filter for which functions apply to this suite
@@ -189,7 +148,7 @@ namespace
 
     TEST_P(aux_test, conversion)
     {
-        RUN_TEST_ON_THREADS_STREAMS(hipblaslt_simple_dispatch<aux_testing>(GetParam()));
+        RUN_TEST_ON_THREADS_STREAMS(aux_testing{}(GetParam()));
     }
     INSTANTIATE_TEST_CATEGORIES(aux_test);