# Changes
1. Modified `hipblaslt-test` and `hipblaslt-bench` to return immediately instead of aborting when no solution is found. This change allows us to check all results, rather than stopping at a single failure, especially useful when using `--yaml` to input multiple test parameters.
2. Removed all unnecessary dependencies on templates from `testing_matmul`. It will now determine the type using the `Argument` parameter.
3. **Removed the type check on `hipblaslt-test` and `hipblaslt-bench` side. It will now output the error message `No solution found`. The main benefit is that there's no longer a need to maintain checks in two places. Also, the output now reveals the actual error from the hipBLASLt API.**
4. Add support for fallback from compute type f16 to f32 on tensilelite

## hipblaslt-bench

### parameters passed via yaml

error.yaml
```yaml
---
include: /home/haosheng/projects/hipBLASLt/clients/include/hipblaslt_common.yaml
Tests:
- name: matmul
  category: quick
  function:
    matmul:
      - { a_type: f32_r, b_type: f32_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r , scale_type: f32_r}
      - { a_type: f16_r, b_type: f16_r, c_type: f16_r, d_type: f16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f16_r, b_type: f16_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf16_r, b_type: bf16_r, c_type: bf16_r, d_type: bf16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf16_r, b_type: bf16_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: f8_ocp_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: bf8_ocp_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: f8_ocp_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: bf8_ocp_r, c_type: f32_r, d_type: f32_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: f8_ocp_r, c_type: f16_r, d_type: f16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: bf8_ocp_r, c_type: f16_r, d_type: f16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: f8_ocp_r, c_type: f16_r, d_type: f16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: bf8_ocp_r, c_type: f16_r, d_type: f16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: f8_ocp_r, c_type: bf16_r, d_type: bf16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: bf8_ocp_r, c_type: bf16_r, d_type: bf16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: bf8_ocp_r, c_type: bf16_r, d_type: bf16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: f8_ocp_r, c_type: bf16_r, d_type: bf16_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: f8_ocp_r, c_type: f8_ocp_r, d_type: f8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: f8_ocp_r, c_type: bf8_ocp_r, d_type: bf8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: bf8_ocp_r, c_type: f8_ocp_r, d_type: f8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: bf8_ocp_r, c_type: bf8_ocp_r, d_type: bf8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: bf8_ocp_r, c_type: f8_ocp_r, d_type: f8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: f8_ocp_r, b_type: bf8_ocp_r, c_type: bf8_ocp_r, d_type: bf8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: f8_ocp_r, c_type: f8_ocp_r, d_type: f8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
      - { a_type: bf8_ocp_r, b_type: f8_ocp_r, c_type: bf8_ocp_r, d_type: bf8_ocp_r, compute_type: c_f16_r, scale_type: f32_r}
  transA_transB:
    - { transA: T, transB: T }
  norm_check: 1
  norm_check_assert: 0
  allclose_check: 1
  initialization: hpl
  timing: 1
  iters: 1
  cold_iters: 0
  print_kernel_info: true
  matrix_size:
    - { M: 1, N: 1, K: 1 }
```

#### original

```diff
/home/haosheng/projects/hipBLASLt+remove_unsupported_transpose_and_datatype/build_25ecb2/debug/clients/staging/hipblaslt_gentest.py --template /home/haosheng/projects/hipBLASLt+remove_unsupported_transpose_and_datatype/build_25ecb2/debug/clients/staging/hipblaslt_template.yaml -o /tmp/hipblaslt-gMoAmn .vscode/error.yaml
[2024-10-21 17:42:35][HIPBLASLT][140724605373931][Api][rocblaslt_create] handle[out]=0x2a57290 
hipBLASLt version: 1000
hipBLASLt git version: 25ecb255
[2024-10-21 17:42:35][HIPBLASLT][140724605373931][Api][rocblaslt_destroy] handle=0x2a57290 
Query device success: there are 1 devices
 -------------------------------------------------------------------------------
Device ID 0 : AMD Radeon Graphics gfx1201
with 17.1 GB memory, max. SCLK 2426 MHz, max. MCLK 1258 MHz, compute capability 12.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
 -------------------------------------------------------------------------------

- Internal error: Test called with invalid types
- function: matmul types:  a: f16_r b: f16_r c: f16_r d: f16_r compute:non-supported compute type
```

#### modified

```diff
hipBLASLt version: 1200
hipBLASLt git version: 1e4ff1f7
Query device success: there are 1 devices
 -------------------------------------------------------------------------------
Device ID 0 : AMD Radeon Graphics gfx1201
with 17.1 GB memory, max. SCLK 2400 MHz, max. MCLK 1258 MHz, compute capability 12.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
 -------------------------------------------------------------------------------
# total 26 * 4 = 104 outputs here ...

+ Is supported 1 / Total solutions: 1
+ [0]:transA,transB,grouped_gemm,batch_count,m,n,k,alpha,lda,stride_a,beta,ldb,stride_b,ldc,stride_c,ldd,stride_d,a_type,b_type,c_type,d_type,compute_type,scaleA,scaleB,scaleC,scaleD,amaxD,activation_type,bias_vector,bias_type,hipblaslt-Gflops,hipblaslt-GB/s,us,CPU-Gflops,CPU-us,norm_error,atol,rtol
+     T,N,0,1,8,8,8,1,8,64,0,8,64,8,64,8,64,bf8_r,f8_r,bf8_r,bf8_r,f16_r,0,0,0,0,0,none,0,f32_r,0.0015139,0.000264361,676.4,0.00039922,2565,nan,1e-05,1e-05
+     --Solution index: 1666
+     --Solution name:  Cijk_Alik_Bljk_B8F8B8S_BH_BiasSHB_HAS_SAB_SCD_SAV_UserArgs_MT64x16x64_MI16x16x1_SN_LDSB0_AFC1_CLR0_EPS0_GRVWA8_GRVWB8_GSU1_GSUC0_GSUWGMRR0_K1_LBSPPA128_LBSPPB128_LPA8_LPB8_LRVW8_MIAV1_MIWT1_1_ONLL0_PGR2_PLR0_SIA3_SS1_SU32_SUM0_SUS256_SVW1_TLDS1_WS32_WG64_2_1_WGM8_WGMXCC1_WGMXCCG0
+     --kernel name:    Cijk_Alik_Bljk_B8F8B8S_BH_BiasSHB_HAS_SAB_SCD_SAV_UserArgs_MT64x16x64_MI16x16x1_SN_LDSB0_AFC1_CLR0_EPS0_GRVWA8_GRVWB8_K1_LBSPPA128_LBSPPB128_LPA8_LPB8_LRVW8_MIAV1_MIWT1_1_ONLL0_PGR2_PLR0_SIA3_SS1_SVW1_TLDS1_WS32_WG64_2_1
+ Is supported 1 / Total solutions: 1
+ [0]:transA,transB,grouped_gemm,batch_count,m,n,k,alpha,lda,stride_a,beta,ldb,stride_b,ldc,stride_c,ldd,stride_d,a_type,b_type,c_type,d_type,compute_type,scaleA,scaleB,scaleC,scaleD,amaxD,activation_type,bias_vector,bias_type,hipblaslt-Gflops,hipblaslt-GB/s,us,CPU-Gflops,CPU-us,norm_error,atol,rtol
+     T,N,0,1,8,8,8,1,8,64,0,8,64,8,64,8,64,bf8_r,f8_r,bf8_r,bf8_r,f32_r,0,0,0,0,0,none,0,f32_r,0.00150677,0.000263116,679.6,0.0144225,71,0,1e-05,1e-05
+     --Solution index: 1666
+     --Solution name:  Cijk_Alik_Bljk_B8F8B8S_BH_BiasSHB_HAS_SAB_SCD_SAV_UserArgs_MT64x16x64_MI16x16x1_SN_LDSB0_AFC1_CLR0_EPS0_GRVWA8_GRVWB8_GSU1_GSUC0_GSUWGMRR0_K1_LBSPPA128_LBSPPB128_LPA8_LPB8_LRVW8_MIAV1_MIWT1_1_ONLL0_PGR2_PLR0_SIA3_SS1_SU32_SUM0_SUS256_SVW1_TLDS1_WS32_WG64_2_1_WGM8_WGMXCC1_WGMXCCG0
+     --kernel name:    Cijk_Alik_Bljk_B8F8B8S_BH_BiasSHB_HAS_SAB_SCD_SAV_UserArgs_MT64x16x64_MI16x16x1_SN_LDSB0_AFC1_CLR0_EPS0_GRVWA8_GRVWB8_K1_LBSPPA128_LBSPPB128_LPA8_LPB8_LRVW8_MIAV1_MIWT1_1_ONLL0_PGR2_PLR0_SIA3_SS1_SVW1_TLDS1_WS32_WG64_2_1
```

### parameters passed via cli (unsupported currently, blocked by hipblas-common [PR #10](https://github.com/ROCm/hipBLAS-common/pull/10))

```diff
[2024-11-15 17:01:01][HIPBLASLT][140720309167462][Api][rocblaslt_create] handle[out]=0x27b9380 
hipBLASLt version: 1200
hipBLASLt git version: 1e4ff1f7
[2024-11-15 17:01:01][HIPBLASLT][140720309167462][Api][rocblaslt_destroy] handle=0x27b9380 
Query device success: there are 1 devices
 -------------------------------------------------------------------------------
Device ID 0 : AMD Radeon Graphics gfx1201
with 17.1 GB memory, max. SCLK 2400 MHz, max. MCLK 1258 MHz, compute capability 12.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
 -------------------------------------------------------------------------------

Invalid value for --compute_type f16_r
```

## rocblas-test

test yaml

```yaml
---
include: rocblas_common.yaml

Tests:
- name: gemm_ex_get_solutions
  category: pre_checkin
  function: gemm_ex_get_solutions
  M: 250
  N: 250
  K: 250
  lda: 250
  ldb: 250
  ldc: 250
  ldd: 250
  alpha_beta: { alpha:  1, beta:  1 }
  transA_transB:
    - { transA: N, transB: N }
  precision :
    - { a_type: f16_r, b_type: f16_r, c_type: f16_r, d_type: f16_r, compute_type: f16_r }
```

#### original

```diff
rocBLAS version: 4.4.0.9c8a7dfe-dirty

rocBLAS-commit-hash: 9c8a7dfeb3d0a808321541567447b5c1d17cd070

Tensile-commit-hash: 09ec3476785198159195e2b8d635db13733682d4

Query device success: there are 1 devices
 -------------------------------------------------------------------------------
Device ID 0 : AMD Radeon Graphics gfx1201
with 17.1 GB memory, max. SCLK 2420 MHz, max. MCLK 1258 MHz, memoryBusWidth 32 Bytes, compute capability 12.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
 -------------------------------------------------------------------------------
info: parsing of test data may take a couple minutes before any test output appears...

/home/haosheng/projects/rocBLAS/build_9c8a7d/debug/clients/staging/rocblas_gentest.py --template /home/haosheng/projects/rocBLAS/build_9c8a7d/debug/clients/staging/rocblas_template.yaml -o /tmp/rocblas-Ikkpj7 .vscode/test.yaml
Note: Google Test filter = *gemm_ex_get_solutions*
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from _/gemm_ex_get_solutions
[2024-10-29 06:19:34][HIPBLASLT][2567709][Api][rocblaslt_create] handle[out]=0x9f7ecf0 
/home/haosheng/projects/rocBLAS/clients/gtest/../include/blas_ex/testing_gemm_ex_get_solutions.hpp:121: Failure
Expected: (size) > (0), actual: 0 vs 0

[2024-10-29 06:19:34][HIPBLASLT][134557032984093][Api][rocblaslt_destroy] handle=0x9f7ecf0 
[  FAILED  ] _/gemm_ex_get_solutions.blas3_tensile/pre_checkin_gemm_ex_get_solutions_f16_rf16_rf16_rf16_rf16_r_NN_250_250_250_1_250_250_1_250_250, where GetParam() = { function: "gemm_ex_get_solutions", name: "gemm_ex_get_solutions", category: "pre_checkin", known_bug_platforms: "", beta: 1.0, stride_a: 62500, stride_b: 62500, stride_c: 62500, stride_d: 62500, M: 250, N: 250, K: 250, lda: 250, ldb: 250, ldc: 250, ldd: 250, a_type: f16_r, b_type: f16_r, c_type: f16_r, d_type: f16_r, compute_type: f16_r, composite_compute_type: invalid, initialization: rand_int, gpu_arch: "", flush_batch_count: 1, transA: 'N', transB: 'N' }
 (283 ms)
[----------] 1 test from _/gemm_ex_get_solutions (283 ms total)

[----------] Global test environment tear-down
[==========] 1 test from 1 test suite ran. (283 ms total)
[  PASSED  ] 0 tests.
- [  FAILED  ] 1 test, listed below:
- [  FAILED  ] _/gemm_ex_get_solutions.blas3_tensile/pre_checkin_gemm_ex_get_solutions_f16_rf16_rf16_rf16_rf16_r_NN_250_250_250_1_250_250_1_250_250, where GetParam() = { function: "gemm_ex_get_solutions", name: "gemm_ex_get_solutions", category: "pre_checkin", known_bug_platforms: "", beta: 1.0, stride_a: 62500, stride_b: 62500, stride_c: 62500, stride_d: 62500, M: 250, N: 250, K: 250, lda: 250, ldb: 250, ldc: 250, ldd: 250, a_type: f16_r, b_type: f16_r, c_type: f16_r, d_type: f16_r, compute_type: f16_r, composite_compute_type: invalid, initialization: rand_int, gpu_arch: "", flush_batch_count: 1, transA: 'N', transB: 'N' }


 1 FAILED TEST
rocBLAS version: 4.4.0.9c8a7dfe-dirty

command line: build_9c8a7d/debug/clients/staging/rocblas-test --gtest_filter=*gemm_ex_get_solutions* --yaml .vscode/test.yaml 

```

#### modified

```diff
rocBLAS info: client (OPENMP) reduced omp_set_num_threads to 62
rocBLAS version: 4.4.0.7c3be2cc-dirty

rocBLAS-commit-hash: 7c3be2cccb8706f84a53cb293f8a7ca023aca8a8

Tensile-commit-hash: 1e25e645f7f60c8edbe55ff1cf789e0daf2468e6

Query device success: there are 1 devices
 -------------------------------------------------------------------------------
Device ID 0 : AMD Radeon Graphics gfx1201
with 17.1 GB memory, max. SCLK 2400 MHz, max. MCLK 1258 MHz, memoryBusWidth 32 Bytes, compute capability 12.0
maxGridDimX 2147483647, sharedMemPerBlock 65.5 KB, maxThreadsPerBlock 1024, warpSize 32
 -------------------------------------------------------------------------------
info: parsing of test data may take a couple minutes before any test output appears...

/home/haosheng/projects/rocBLAS/build_7c3be2/debug/clients/staging/rocblas_gentest.py --template /home/haosheng/projects/rocBLAS/build_7c3be2/debug/clients/staging/rocblas_template.yaml -o /tmp/rocblas-mWigS6 .vscode/test.yaml
[==========] Running 1 test from 1 test suite.
[----------] Global test environment set-up.
[----------] 1 test from _/gemm_ex_get_solutions
[2024-11-15 17:09:42][HIPBLASLT][710463][Api][rocblaslt_create] handle[out]=0x9e2db10 
[2024-11-15 17:09:43][HIPBLASLT][710463][Api][getAllSolutions] Found hardware solutions: =53 
[2024-11-15 17:09:43][HIPBLASLT][710463][Api][getAllSolutions] Final hardware solutions: =53 
# ...
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Found hardware solutions: =53 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Final hardware solutions: =53 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Found hardware solutions: =53 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Final hardware solutions: =53 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Found hardware solutions: =53 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Final hardware solutions: =53
# ...
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][rocblaslt_matmul_get_algos_from_index_cpp] returnAlogCount=1 
# ...
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][rocblaslt_algo_get_heuristic_cpp] returnAlogCount=1 
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][rocblaslt_matmul_get_algos_from_index_cpp] returnAlogCount=1 
# ...
[2024-11-15 17:09:44][HIPBLASLT][710463][Api][getAllSolutions] Final hardware solutions: =58 
[2024-11-15 17:09:44][HIPBLASLT][4605390717358626623][Api][rocblaslt_destroy] handle=0x9e2db10 
[----------] 1 test from _/gemm_ex_get_solutions (2187 ms total)

[----------] Global test environment tear-down
+ [==========] 1 test from 1 test suite ran. (2188 ms total)
+ [  PASSED  ] 1 test.
rocBLAS version: 4.4.0.7c3be2cc-dirty

command line: build_7c3be2/debug/clients/staging/rocblas-test --yaml .vscode/test.yaml 
```