@cliffxzx 2024-10-21
MI200

```
[----------] Global test environment tear-down
[==========] 13061 tests from 13 test suites ran. (939330 ms total)
[  PASSED  ] 13061 tests.
hipBLASLt version: 1100

command line: build_1b0889/debug/clients/staging/hipblaslt-test 
```