@cliffxzx 2024-11-17
gfx1201

```
[----------] Global test environment tear-down
[==========] 34847 tests from 13 test suites ran. (731100 ms total)
[  PASSED  ] 33963 tests.
[  FAILED  ] 884 tests, listed below:
# ... all 884 failed tests are due to a known bug related to bf8..
884 FAILED TESTS
hipBLASLt version: 1200

command line: build_576d0c/debug/clients/staging/hipblaslt-test 

```